Remember Skipped Rows In Import
This is needed to select unused data in the import process later after filtering (for null values for example).

To be honest; I do not think this is the best way to do that, but is the fastest for now.
I would prefer seperating the business logic into an no-vue typescript only class or whatever, and this I should have done from the beginning.
But I didn't, and instead I

Changes:
- modified:   datacontainers.ts:	Add select rows by index method.
- modified:   dataconversion.ts:	Seperate missing required fields logic, new getMissingRequiredFieldIndexes function
- modified:   LemmaFormatter.vue:	Pass missing Row indexes.
- modified:   LemmaImporter.vue:	Provide raw imported data of the rows, that will not be skipped later on.
- modified:   NullManager.vue:		Provide the indexes ov rows, that have to be skipped.

Tags: [skip ci]